requirements.txt
streamlit
beautifulsoup4==4.11.1
bs4==0.0.1
pip==23.1.2
requests==2.28.0